Properly remove / apply effects during reattach.
Previously, when an attachment was re-attached from one card to another,
it would leave its effect on the original parent in addition to applying
it to the new parent. Now, effects are removed immediately when the
parent changes, and effects are applied to the new parent asynchronously
after the move has completed.