Handle characters with null strength.
Per this change to the JSON data format, cards that have X strength like
Dacey Mormont and Beric Dondarrian are represented with a NULL strength.
https://github.com/Alsciende/thronesdb-json-data/pull/160/commits/b5982faf8126c4f030af26dc71d28150f235983b

Such cards were not displaying their strength at all because the
serializer was always sending strength 0 even if the strength was
boosted through effects. Now, the printed strength and current strength
are fully calculated regardless of whether the JSON field is null. This
should ensure strength is always displayed properly. It should not
affect location and attachment cards, as their strength was already
being sent as 0 and `DrawCard.getStrength` now always returns 0 if
strength is not defined.