Ensure take control adds card to `allCards` array.
Previously, cards that were taken control by putting them directly into
play under the new controller (e.g. marshalling via Night Gathers) did
not add the card to the `allCards` array for that player. This lead to
bugs where the card was clearly in play, but checks that used
`filterCardsInPlay` would fail. Now, the `putIntoPlay` method moves the
card to the `allCards` array the same way that the `takeControl` method
previously did.