Fix ordering of leaves play events / effects. (#1108)
Previously, when a card would leave play it would remove attachments and
dupes from the card prior to interrupts firing. This caused a bug where
Marriage Pact's forced interrupt would not fire because it was detached
from the parent before the parent started to leave play.

Now, all interrupts fire for the card leaving play, then attachments and
dupes on the card leave play, then reactions to the card leaving play
fire.