Trigger reaction events for Annals of Castle Black. (#1077)
Previously, Annals of Castle Black would not allow event interrupts /
reactions to trigger from discard. This is because event listeners were
only ever bound if the card was in the expected location - in the case
of event cards, in hand.

Now, interrupts / reactions listen to events in all open information
locations plus while in hand. Since the `location` property restricts
whether the ability actually fires, abilities should not activate while
in inappropriate locations even if it listens to events.