Save cards for ability activator for Iron Mines.
Because Iron Mines sacrifices itself as a cost to activate, it is placed
in its owner's discard pile before the target for the save is prompted.
This means that the controller has reverted to the original owner by the
time targeting happens. This caused a bug where a taken-control-of Iron
Mines would trigger if any of your characters were killed, but then
would only allow you to use it to save your opponent's characters.

Now, Iron Mines explicitly compares the target's controller with the
player that activated the ability.