Discard active plots before flipping new ones.
Previously, the active plot was placed in the used pile in sequence with
revealing the new one. This could lead to situations where if the same
plot was revealed twice in a row, such as Retaliation, the unapply of
original's effects could happen after the apply of the new effects, thus
undoing it.

Now, each player's active plots are moved to the plot discard and
deactivated prior to revealing the new plots.