Remove "winner determined" action window
The "winner determined" action window was put into place to allow a
pause to play in-hand event cards. Now that nearly all events in hand
prompt the player directly, the window is no longer needed.
Additionally, the presence of the window leads newer players to believe
they can ambush cards or play Action events in this pause, which is not
allowed by the rules.