Fix prompts for cards with maximum limits. (#1181)
Previously, if a copy of a card with a maximum limit was in the discard
pile, the player would not be prompted to trigger that card if they had
another copy in hand. This was because the copy in discard would
register with the prompt first, preventing any other copies from being
registered. Then, when the prompt activated it saw that the registered
copy was in discard and therefore couldn't be triggered.

Instead of preventing multiple copies from being registered, it is now
filtered when the prompt is generated and presented to the player.