Remove plots before revealing new ones (#1183)
* Use plot discard event for Rains of Castamere.

Instead of removing the plot manually when plots are flipped, move the
card out of the game when it enters the plot discard.

* Only select plots from the plot deck for tests.

Previously it could reach into the used plot pile and replay the same
plot.

* Discard active plots before flipping new ones.

Previously, the active plot was placed in the used pile in sequence with
revealing the new one. This could lead to situations where if the same
plot was revealed twice in a row, such as Retaliation, the unapply of
original's effects could happen after the apply of the new effects, thus
undoing it.

Now, each player's active plots are moved to the plot discard and
deactivated prior to revealing the new plots.