Remove crashing abilities from prompts
Previously, if the eligibility check for an ability throw an exception,
the entire prompt would crash, leaving players with a prompt with no
buttons. Now, if an ability throws an exception when checking
eligibility, the error will be reported and then the ability will simply
be removed from the prompt.