Better crash recovery (#1211)
* Cancel ability after error during resolution

Previously, if an ability threw an error it could reach an unrecoverable
state such that it wouldn't clear the game pipeline. Now, if an error
occurs at any step of resolution (e.g. targeting, paying costs, the
handler, etc) then the entire ability resolution will be cancelled.

* Remove crashing abilities from prompts

Previously, if the eligibility check for an ability throw an exception,
the entire prompt would crash, leaving players with a prompt with no
buttons. Now, if an ability throws an exception when checking
eligibility, the error will be reported and then the ability will simply
be removed from the prompt.