Recalculate effects when traits change
Previously, if an effect were already in play that was limited to cards
with a certain trait, and a card already in play gained or lost that
trait, the effect would not be applied or removed from that card. For
example, characters that gained the 'Knight' trait while Tourney for the
King was revealed did not gain the renown keyword.

Now, when a card gains or loses a trait, effects for that card are
recalculated in a similar manner to the way they are recalculated for
faction changes.