Calculate Crossing STR based on challenge number (#1235)
* Record challenge number on Challenge object
* Calculate Crossing STR based on challenge number

Previously, Lord of the Crossing's STR bonuses were calculated based on
the number of challenges initiated by the player. This lead to a bug
where after attackers were picked but before the challenge was
initiated, the attackers would get the -1 STR penalty. If the opponent
had Blood of the Dragon at the time, this could kill the attackers.

Now, the challenge number is calculated up from and remains constant for
the entire challenge, preventing such bugs.

* Convert other cards to use challenge number