Revert controlled cards when they move (#1261)
Previously, if a card that had been taken control of was sacrificed or
returned to hand, it would properly update the `controller` property.
However, the card would not be removed from the list of cards controlled
by the opponent nor added back to the list of cards controlled by the
player. This caused bugs where a controlled character leaving play could
not be controlled again if they re-entered play somehow.