Add `chooseOpponent` option for abilities (#1283)
* Add prompt to select an opponent

* Rethrow original error in tests

* Add 'chooseOpponent' option to abilities

Adds a `chooseOpponent` option to abilities for cards that explicitly
ask to choose an opponent. The option can either be a function that
takes the opponent object and returns whether they're allowed to be
chosen, or the value `true` for abilities that allow any opponent to be
chosen.

* Convert explicit "choose an opponent" cards

Also fixes a bug where Night Gathers was allowed to be used while not
having a higher reserve than the opponent.

* Convert cards that implicitly choose an opponent

* Allow player to cancel choose opponent prompt

* Check opponent requirements upfront

When checking whether an ability meets requirements, it now checks
whether there are any eligible opponents to be chosen if the ability
requires it.