Convert reserve discard into a prompt
Previously, the prompt that checked that players were below reserve
expected them to manually drag cards from hand into their discard pile.
This was inconsistent with the rest of the site, error prone (easy to
drag onto draw pile instead), and made things difficult for browsers
with finicky drag support (Firefox, touch-based dragging).

Now, when the player is above their reserve value, they will be prompt
to select the cards they wish to discard. This prompt maintains the
order of the cards selected so that the player can control the order the
cards end up in discard.