Allow custom cards in the deck builder
Custom cards can now be added to a deck by manually adding a line to the
deck in the format "Custom <Type> - <Name>". For example:

2 Custom Character - Ser Pounce

will add 2 copies of a character named Ser Pounce to the deck.

Such custom cards will be blank, neutral, traitless, and have 0 cost and
0 strength. The stats should be manipulated via in game commands to
match the intended card.