Allow drag-drop from additional card piles (#1302)
* Pass card directly into Player.drop method

* Restrict drag targets by type instead of source

Because the out of game pile can contain both draw and plot cards,
players should be able to drag cards from there to both draw and plot
cards. However, they should not be able to drag draw cards to plot piles
or vice versa. Thus, dragging now checks whether the card type is being
dragged into a pile that allows that type instead of just checking the
source and target location.

* Don't crash if the pile has no cards

* Add proper sources to piles for dragging

Adds the source name for the pile and the onDragDrop handler for the
optional card piles including out of game, Rains schemes and Conclave
cards.