Reset lasting effects when card leaves play
Previously, only persistent effects were removed when a card left the
play area. This lead to bugs where a card that left play but entered
play again during the period for a lasting effect would continue to have
that effect. With burn effects, this would lead to the card re-entering
play with the strength penalty still in place.