Override min reserve when manually setting reserve (#1327)
Previously, if a minimum reserve was set by something like Wraiths in
their Midst, it was impossible to go below that reserve even when
setting it manually. Now, any adjustment to reserve will get rid of the
minimum requirement. This behavior is more in line with other overrides
and should help players get past the NaN reserve bug while we track it
down.