Set correct cardtype for event+attach saves (#1363)
Previously, after playing and attaching a Risen from the Sea/Blood Magic Ritual to a character, these cards still retained their event cardtype. This made it for example impossible to discard these cards with Confiscation. This PR fixes that by setting the correct cardtype as long as these cards remain in play.