Allow multiple reactions from a card in a window (#1380)
Previously, it was assumed that if a prompt had multiple choices from
the same card in a single window that it was because the card offered
multiple choices (e.g. The Reader, LoCR Tyrion). Thus, if the player
clicked one of the options, the others were removed.

However, now that the challenge initiation window contains choices for
multiple events, it's possible to have the same card produce multiple
independent choices. For example, Marya Seaworth can trigger separately
for each card bypassed by stealth. The previous implementation only
allowed one instance of her ability to be triggered. The prompt now
treats them as separate abilities instead of separate choices for the
same ability, and allows all of them to be triggered.