Activate effects at the end of setup phase
Previously, effects of cards placed during setup were immediately
active, even if not immediately visible to the player. This led to bugs
such as Elinor Tyrell allowing an extra limited card to be played during
setup.

Now, when an effect is created, its `active` property is determined by
whether or not the source of the effect is facedown. Thus, cards placed
in setup will not be active immediately. At the end of setup, all
persistent effects are explicitly activated.