Update cards to use movePower
Previously, some cards were using two separate calls to modifyPower to
subtract power from one card and add it to another. This lead to bugs
where if the source card was immune to an effect, its power would not be
subtracted but it would be added to the target card.