Distinguish between gaining and moving power (#1475)
* Generalize and rename transferPower to movePower

The transferPower method has been renamed to movePower and now will work
with any type of card, not just faction cards. The parameter order has
been reverse to have the source card come first followed by the target
card to clarify this change.

* Update cards to use movePower

Previously, some cards were using two separate calls to modifyPower to
subtract power from one card and add it to another. This lead to bugs
where if the source card was immune to an effect, its power would not be
subtracted but it would be added to the target card.

* Enforce immunity for moving power

* Distinguish between power moved and gained

* Change onCardPowerChanged to onCardPowerGained

The event will now only be raised when a card gains power per the game's
definition of 'gain.'

* Convert Edmure Tully to use target API

Technically Edmure doesn't target, but it's more convenient to use the
target API here.

* Update prompt titles to use interpolation