Characters killed by burn maintain their strength
Per the ruling on Quentyn Martell, characters killed by burn are killed
prior to having their strength modified. So if Quentyn is killed via
Dracarys!, his STR during his interrupt ability is 4. Thus, instead of
having the burn effect kill when it detects a 0 STR character, the
modifyStrength function now checks this directly in case the STR would
be reduced to 0.