The Sordid Tale of Quentyn Martell (#1486)
* Get tests in place

* Update burn to be a single effect

Instead of applying burn as two separate effects - a killByStrength flag
and the strength modifier - they are now combined into a single effect.
This should be less error prone in the future and allow ordering of the
overall effect to be changed if necessary.

* Characters killed by burn maintain their strength

Per the ruling on Quentyn Martell, characters killed by burn are killed
prior to having their strength modified. So if Quentyn is killed via
Dracarys!, his STR during his interrupt ability is 4. Thus, instead of
having the burn effect kill when it detects a 0 STR character, the
modifyStrength function now checks this directly in case the STR would
be reduced to 0.