Do not duplicate controlled cards in allCards
Previously, the removeCardFromPile method was inappropriately adding the
passed card to the allCards array while attempting to revert a
controlled card back to its owner. This lead to controlled cards
appearing in the allCards array, which could cause problems when
filtering cards, such as Varys discarding the same card multiple times
and having it end up in the wrong pile.