Fix Rains + Calm + Riddle interaction (#1499)
Varys' Riddle initiates Calm Over Westeros' "When Revealed" ability as
if that player had revealed it. The lasting effect of Calm lasts until
the revealing player (the Riddle player) reveals a new plot, not until
the current controller reveals a new plot. Thus, the Calm player
revealing a new plot via Rains of Castamere should not cancel Riddle's
copy of the lasting effect.