Fix improper discards of controlled cards (#1498)
* Do not duplicate controlled cards in allCards

Previously, the removeCardFromPile method was inappropriately adding the
passed card to the allCards array while attempting to revert a
controlled card back to its owner. This lead to controlled cards
appearing in the allCards array, which could cause problems when
filtering cards, such as Varys discarding the same card multiple times
and having it end up in the wrong pile.

* Never move card out of play to non-owner

Previously, moveCard would properly redirect a card moving from the play
area to an out-of-play area to its owner's pile. However, a card moving
between two out-of-play areas (e.g. hand to discard, discard to itself)
allowed non-owned cards to be moved.