Apply plot persistent effects simultaneously
Instead of applying plot effects one at a time as they are flipped (in
player-join order), they are now applied simultaneously per the rules.
This fixes a bug where if the player that created the game revealed
Blood of the Dragon and the opponent revealed A Song of Summer, the
opponent's 1 STR characters would improperly die.