Refresh user auth when blocklist is modified (#1509)
In order to update the blocklist on both the client side JWT and server
side user objects, a new token is sent back with the updated blocklist,
similar to what is done on profile updates. Then, the client reauths
with the server using the token, which updates the user object on the
corresponding socket on the server side.