Add 'game element type' for use with effects
Previously, the effects engine would check whether a potential target
was an instance of a specific class when checking the `targetType`
property. This was inflexible, leading the Effect class to need to know
more about what it was being applied to than necessary, including
creating actual instances in unit tests.

Now, each class that can receive effects must implement a
`getGameElementType` method that returns a string. This string is
directly compared against the `targetType` property when applying
effects.