Fix NaN reserve bug (#1542)
Previously, the setMinReserve effect stored the previous value for the
player's min reserve using the player ID as a key. This lead to a bug
where if a player who was affected by Wraiths in their Midst reconnected
after it was revealed but before the next plot was revealed, their ID
would change. Thus, when the effect was unapplied, the stored value
could not be retrieved because it was under the original ID.

Now, the original min reserve is stored under the player's name, which
should not be able to change mid-game.