Refactor costs (#1545)
* Extract ability contexts to separate class

* Implement cost builder for card-targeting costs

Most costs are applied to a particular card - either itself, its parent,
cards selected by the player, the faction card, or a specific card in
response to an event. This concept has been extracted into a CostBuilder
class that separates the logic of what the cost does (e.g. kneel,
sacrifice) from which card is affected.

Additionally, the selections for costs are now uniformly stored on the
`context.costs` object under the name of the action applied to the card.
For example, a cost that kneels a card will store that card as
`context.costs.kneel`.

* Implement cost actions and builder

Extracts each action for a cost (e.g. kneel, sacrifice) into its own
class to be passed into CostBuilder. Also adds CostBuilders which
contains builders for each type of action.

* Convert kneel costs

* Convert sacrifice costs

* Convert discard from hand costs

* Convert discard power costs

* Convert return to hand costs

* Convert remove from challenge costs

* Convert kill costs

* Convert put into play costs

* Convert remove from game costs

* Convert reveal card costs

* Convert stand costs