Do not place events in discard immediately (#1549)
When an event card is played, it goes into a temporary "being played"
pile to indicate that it is currently resolving. Once the resolution of
the ability has completed, it is then placed in discard.

From a UI perspective, cards that are currently being played are
displayed as if they are already in the discard pile.