Implement Emissary of the Hightower (#1555)
* Rework playable locations to take a predicate

Previously, playable locations was set up to allow you to take a
specific play type action from a particular player location (e.g.
marshal from opponent's discard).

Now, playable locations take a predicate that can be used more
generally. This is necessary in order to support Emissary of the
Hightower, which allows one specific card to be marked as playable from
the discard pile, instead of marking the whole discard pile playable.

* Allow cards in discard to be targeted by effects

Emissary of the Hightower requires that an effect be applied to a
specific card in the discard pile. Thus, the effects engine needs to
allow it as a valid target.

* Implement Emissary of the Hightower