Record costs in array to prevent overwriting
The previous cost refactor introduced a bug where if a card used the
same cost action on two different cards (e.g. Joffrey (GoH) kneeling the
loyal card and the faction card), only one of the cards would get the
cost action applied. This was due to the entry in the costs object being
overwritten, whereas previously each cost was written manually and
didn't always store an entry on the costs object.

Now, when a cost choice is made, its recorded to a separate parallel
object that always stores the costs as an array. The previous mechanism
is left in place for compatibility with abilities that need to look up
the exact card (typically for the message in chat).