Do not allow removal of non-existent tokens (#1566)
Previously, if an ability attempted to remove a token from a card that
didn't have that token, it would set the token value to NaN. This
displayed the token in the UI but did not count as having the token for
the purpose of the `hasToken` method. The main scenario involved is
Jaqen places a Valar token on a character. That character leaves and
re-enters play. Then Jaqen leaves play and tries to remove the original
token placed, which was already removed when the targeted character left
play.