Only fire onDeclaredAsDefender for chosen cards
Previously, the onDeclaredAsDefender event was being fired for all
defenders within a challenge. However, it should only fire for cards
specifically chosen / forced to be chosen by the player. This lead to a
bug where cards added to the challenge by an ability (e.g. HoT Margaery)
would be declared as a defender even though they weren't actually
declared.