Handle setting multiple challenge limits at once
Previously, if two copies of Besieged were revealed at the same time,
the defender minimum would not reset properly after the next set of
plots were revealed. This was due to the second copy of Besieged seeing
the value set by the first one, and reseting it to that value. The bug
did not occur with double Jousting Contest because the maximum was being
explicitly reset at the start of the plot phase.

Now, both mins and maxs are stored on an array. Once all values are
unset, the default value will be returned again instead of overwriting
values.