Fix webpack build scripts (#1621)
* Fix webpack build scripts

Adds a `build-vendor:dev` script for local development. This fixes the
issue where building the vendor scripts in production mode but running
the server in development mode caused ReactDOMServer.renderToString to
crash.

Removes the `-p` option from the build scripts. This was conflicting
with the usage of UglifyJsPlugin, basically uglifying Javascript code
twice. See:
https://github.com/webpack/webpack/issues/1385#issuecomment-160553651

Removes the unnecessary passing of environment variables to Webpack.
Environment is now controlled completely using the `--env` option.

* Remove colon from build-vendor-dev

* Incorporate offline feedback re: webpack cli options