Fix abilities related to cards placed in discard
Per the FAQ, abilities that interact with a card while in an out-of-play
area require that it be in that area in order to trigger. For example,
when a card is placed in the discard pile but is then moved via Now My
Watch Begins, the ability on Nagga's Ribs can no longer be triggered
because the card is no longer in discard.

See more:
http://thronesdb.com/faq##Card_Abilities_and_Out_of_Play_Areas