Filter out cards missing card data from decks
Previously, if a single deck in the deck list contained a card whose
code was not in the list of available cards, the entire deck list would
refuse to load. This situation likely only occurred in development mode,
after creating a deck with a card then reseting card data to a data set
that did not include that card. Now, such cards are simply filtered out
of the deck prior to validation.