Fixes to Pay The Iron Price / attachment control (#1700)
* Fix control of attachments for Pay The Iron Price

Previously, when Pay The Iron Price took control of an attachment, the
attachment would end up in two places in the game board - once on the
character, once in the opponents discard pile. This made it impossible
to discard the attachment as well as other problems.

* Pass function for take control effect / Ward

Previously, the player that gained control of a character attached with
Ward was fixed when the card was created. With Pay The Iron Price, it
needed to be the current controller of Ward at the time it was attached.