Fix "does not kneel when declared as attacker"
The previous change to the effects engine where effects are recalculated
only after a game event has occurred caused the "does not kneel when
declared as an attacker" cards to stop working. This is because the
event for attackers being declared and being knelt happens
simultaneously, so the engine had no opportunity to see that the cards
shouldn't be knelt until after kneeling them.