Do not discard attachments that are being removed (#1721)
The addition of the attachment validity check lead to a bug where
non-terminal attachments did not return to hand properly. This was due
to the fact that by the time the asynchronous return to hand call
occurs, the parent of the attachment is already out of the play area,
which then triggers the attachment validity check to discard it instead.