Do not recalculate effects that have expired (#1725)
With the previous change to the effects engine, conditional and state
dependent effects are queued up to be recalculated after each event.
However, certain events around revealing new plots cause those effects
to be removed. Since the effects were already queued, though, the engine
ended up reapplying the effect after it expired.

Now, the engine makes sure the effect is still in the list of active
effects before reapplying it.