Automatically recalculate player-matched effects
Previously, when an effect was applied only to players that met a
matching condition (e.g. Tommen, The King in the North, A Game of
Thrones), an explicit recalculate was needed. The change to the way
effects are recalculated didn't take this into account, requiring an
always-true condition be added to such effects to ensure they were
recalculated. If that condition was missing, the effect would never be
reapplied properly.

This led to a bug where A Game of Thrones would not allow players
initiate military or power challenges, even after winning intrigue
challenges, because the effect wasn't removed from the player.

Now, all effects that target a player that also have a matching function
will automatically be considered a conditional effect and be
recalculated at the proper interval.