Don't show warning message for Nothing Burns when no valid locations (#1750)
Previously, the target prompt for Nothing Burns Like the Cold precluded selecting locations that can't be discarded, but the hasValidLocations() check did not. This could lead to players being presented with a prompt with all targets greyed out, but after cancelling getting a message warning players they did have valid targets. This PR fixes that.