Use ability contexts directly for immunity
Previously, immunity / cannot restrictions looked at an object that
contained the card and the current stage of resolution for the ability.
This was referred to as a "ability context" which overlapped with the
name given to the context that tracks the player using the ability and
cost / target resolution.

To reduce confusion and gain access to this additional information, the
actual AbilityContext object is tracked directly and used within the
immunity system.