Introduce additional gold sources
Adds an effect that allows gold to be spent from a source other than the
player's gold pool. When the player goes to spend money, they will be
prompted how much from the separate sources they want to spend. If the
amount of gold they are spending can only be covered by all gold from
all sources, or if only one source has enough gold to cover it, then
they will not be prompted and the gold will be automatically spent.