Better encapsulate gold usages (#1807)
* Add modifyGold method to cards and players

Unifies the interface for modifying the amount of gold for both player's
gold pools and on cards.

* Encapsulate returning player gold to treasury

* Allow transferGold to work with cards

Previously, the transferGold method only worked for moving gold between
two player objects. Now, it may be used to transfer gold between any
combination of players and card by using the unified modifyGold method
internally.

Additionally, the interface for the method has been updated to take an
object parameter. By forcing the caller to explicitly specify the `from`
gold source and `to` gold destination, the usage of the method should be
more readable than the previous 3 parameter version.

* Encapsulate spending gold