Fix ability limit when card leaves play as cost
Previously, if an ability had both a cost that caused the card to leave
and a limit, the limit would be improperly incremented after the card
left play. This could cause cards like Obara Sand (SoD) to only be
usable once per phase, despite the rules saying the limit should reset
when they leave / re-enter play.