Only allow Forced March to initiate twice
The post-then effect of Forced March reads as "initiate this effect
again." The "this effect" refers to the pre-then effect, not to the
ability as a whole. This means initiating the post-then effect does not
lead to another instance of the "then", and thus you can only initiate
the effect a total of two times.