Convert ability timer to use redux (#1933)
* Move ability timer state into Redux

The state surrounding the ability timer has been moved into Redux in
order to allow the timer to be cancelled from outside the
ActivePlayerPrompt component.

Additionally, the timer now renders at the same render cycle as the
browser by using `requestAnimationFrame`. This should help ensure
smoother rendering than calls previously made by `setInterval`.

* Activate ability timer from redux

Previously, the ability timer was activated when the ActivePlayerPrompt
component was rendered containing a timer button. This could lead to
situations where the timer is cancelled outside the prompt, which caused
the prompt to rerender with the blank timer, which then caused it to
re-start the timer.

* Stop ability timer on card click / menus