Add nested, attached events w/ separate interrupts
This allows child events added from an event handler to fire their own
separate interrupt window. As an example, kill via burn is nested inside
of a strength reduction event - interrupts to the STR reduction occur
first, the STR reduction is applied, then interrupts to the kill occur,
then the kill is applied, then reactions to both the STR reduction and
kill occur.