Clear ability resolution when leaving play area
If a card whose ability would trigger for a specific triggering
condition leaves the playable area and re-enters the playable area while
that triggering condition is still active, the player should be able to
trigger that ability again.

For example, if you have an event in hand whose condition has been met,
play that event, then return that event back to hand through a nested
sequence (Maester at Rock to put the event on the draw deck, Lannisport
to re-draw the event), you should then be able to play that same event
again.