Fix timing of "When Revealed" copying abilities
"When Revealed" copying abilities like Varys's Riddle, Pulling the
Strings, and Tyrion's Chain now directly resolve the ability rather than
raising an event. This is because the resolution of the copied ability
is considered part of the resolution of the copier's ability, rather
than a nested sequence. As such, the copied ability directly uses the
same event object received by the copier's ability.