Fix timing for At Prince Doran's Behest (#1957)
* Add nested, attached events w/ separate interrupts

This allows child events added from an event handler to fire their own
separate interrupt window. As an example, kill via burn is nested inside
of a strength reduction event - interrupts to the STR reduction occur
first, the STR reduction is applied, then interrupts to the kill occur,
then the kill is applied, then reactions to both the STR reduction and
kill occur.

* Fix timing on At Prince Doran's Behest

* Fix timing of "When Revealed" copying abilities

"When Revealed" copying abilities like Varys's Riddle, Pulling the
Strings, and Tyrion's Chain now directly resolve the ability rather than
raising an event. This is because the resolution of the copied ability
is considered part of the resolution of the copier's ability, rather
than a nested sequence. As such, the copied ability directly uses the
same event object received by the copier's ability.