Click cards to trigger interrupt / reaction abilities (#1934)
* Move choice selection to separate prompt

* Allow choosing target card for multiple events

When the same card can respond to multiple triggering conditions
simultaneously, the player now clicks on the card that raised the
condition directly rather than clicking a button.

For example, if multiple characters are bypassed by stealth, it should
trigger multiple instances of Marya Seaworth's ability. Previously, the
player would click 'Marya Seaworth - Kneel Character 1' to trigger the
specificly targeted character. Now, the player clicks 'Marya Seaworth'
followed by directly clicking on the targeted character.

* Click cards to trigger abilities

* Make top card of piles clickable in select prompts

If the top card of a pile is selectable for a card select prompt, and no
cards inside of the pile are clickable, clicking on the pile will now
count as clicking on the top card instead of opening the pile.

* Disable menus on agendas during select prompts

* Fix timer during card select prompts