Ensure plot revealed by Behest triggers
When At Prince Doran's Behest was revealed as the only plot with a 'when
revealed' ability, the ability window was closing immediately after
executing Behest. This lead to the plot revealed by Behest not
executing, because the window it would be executed in was already
closed.

Now, the 'when revealed' ability window will stay open until it has done
one final cycle and verified there are no more abilities needing to be
executed.