Only apply attachment effects if there's a parent
Previously, attachment related `whileAttached` effects were always
active, using a matching function to limit its targets to only the
parent card (if there was one). This led to crashes when the effect was
dependent on a parent card being present, such as Tokar. Now, the
attachment must have a parent before the effect is applied, and losing
the parent will immediately unapply the effect entirely instead of
recalculating.

Fixes THRONETEKI-26R