kill costs can be payed only using in-play characters (#1991)
Currently Given to the Drowned God allows to use in-hand cards to pay its
cost. While that could be fixed in the card, as per the rules killed characters
"leave play", so they have to be in play to be killed. Hence this change, which
enforces the fact that kill costs can be played only using in-play cards.