Do not show opponents the card during prompt
Previously, if a card had multiple play actions (such as Lady in
Waiting), the opponent was able to see what card was being used while
the player chose which option they wanted. This would reveal which card
a player is considering to put into shadows, so showing the card needs
to be disabled for that prompt.