Remove restrictions when blanking an attachment
Previously, blanking an attachment did not blank the attachment
requirements imposed by the card text (e.g. only certain factions, only
opponent cards, etc). Normally this doesn't make a difference, but if an
attachment specifies that it goes on a location, blanking should cause
it to revert to an attachment for characters, which should cause the
attachment to be discarded as invalid.