Treat Ambush, Bestow, Shadow and No Attachments as normal keywords (#2025)
* Convert ambush to normal keyword
* Convert bestow to normal keyword
* Convert shadow to normal keyword
* Convert no attachments to normal keyword

* Extract keyword costs to a new property class
* Fix Blessed Maiden
* Fix attachment trait logic and clarify naming

If a card somehow gains multiple "No attachments" keywords, all
instances of that keyword must be satisfied for an attachment, not just
a single instance. For example, a card with "No attachments" and "No
attachments except Weapon" should not allow any attachments. Similarly,
a card with "No attachments except Item" and "No attachments except
Weapon" only allows attachments with both an Item and Weapon trait.