Implement saves using singular events (#2029)
* Implement saves using singular events

Previously, we used fake plural events such as onCharactersKilled to
simulate simultaneous resolution. Since then, support has been added for
simultaneous events as well as better UI for choosing which event you
want to match when triggering an ability. It is now possible to
implement saves directly on the singular onCharacterKilled event.