Support functions for `disabled` button prop
Allows the `disabled` property on buttons to be a function. This allows
the disabled state of a button to be updated in response to other game
events. For example, if the challenges prompt has challenge types that
are disabled, the /reset-challenges-count command will re-enable those
buttons without having to closer and reopen the prompt.