Event system enhancements (#2038)
* Allow post-handlers to be chained on events

* Copy event properties onto atomic events

Because atomic events represent a simultaneous execution of multiple
events, it is useful to be able to access the properties of those events
directly instead of needing to iterate over the child events directly.

* Fix ordering of attached events

Attached events should resolve before the post-handler functions
associated with an event. Thus, the interrupt window needs to open
immediately after the handler for the event.