Fix how allowed challenges are tracked (#2057)
* Only use allowed challenges for attacking player

* Don't remove effects when reseting challenges

When `ChallengeTracker.reset` was being called at the start of the
round, it was clearing out the list of allowed challenges. This lead to
a bug where if Khal Drogo was setup or put into play the previous round,
during the current round his additional military challenge would be
gone.

Now, instead of removing allowed challenges, they're simply marked as
used until the next time `ChallengeTracker.reset` is called.