Convert challenge effects from bool to counted (#2058)
Previously, challenge-related effects such as "does not kneel as
attacker / defender" were implemented by setting and unsetting booleans.
This lead to bugs where if the same card received multiple copies of the
effect (e.g. attaching The Wolf King to The Blackfish at 4 power), when
one of the effects was unset it would erase the other effect.

Now these effects are tracked using a reference-counted set similar to
traits and keywords, to ensure that removing one effect won't remove all
of them at once.