Check for unserializable data in integration tests (#2084)
Since integration tests do not run an actual game server + socket.io,
the tests would not detect changes that added unserializable data to
either game messages or as properties on player or card state. Now the
integration tests check that there is no unserializable data in the game
state that would be sent to the client after every player input in the
tests.