Track which targets have effects applied
Previously, if a target for an effect had immunity to that effect, it
simply wasn't added as a target. In order to handle cases where immunity
wears off after the effect is already applied, it will now be added as a
target but not have the effect applied to it.

Additionally, gaining immunity after a lasting effect has already been
applied will no longer unapply the effect, per RRG rules.