Ensure first player order for reserve prompt (#2114)
Previously, the prompt to discard down to reserve was being created at
the beginning of each round. Because of this, the `remainingPlayers`
array was always in first player order for the previous round, which may
not match the first player order for the current round after plots are
revealed.

Now the prompt is created on demand when the proper step in the pipeline
has been reached.

Additionally, a message has been added when a player does not have to
discard to meet reserve to make it more clear for both players and
spectators on what's happening.