Add support for card counts in test deck builder
Previously, if you wanted to add a specific number of copies of a single
card to a deck in a test, you would have to list that card's name that
many times in the call to `buildDeck`. Now there's a short hand
available that allows you to specify the count directly.

Example:
```javascript
this.buildDeck('stark', [
  'Trading with the Pentoshi', 'A Noble Cause',
  'Nightmares', 'Heart Tree Grove', 'Needle',
  { name: 'Arya Stark (Core)', count: 3 }
]);
```